pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend="hatchling.build"

[project]
name = "statsized"
authors = [{name = "Statsized", email = "capsizedhockey@gmail.com"}]
description = "Stat streaming from NHL.com"
readme = "README.md"
license = { text = "Apache-2.0" }
version = "0.1.0"
requires-python = ">=3.9"
keywords = []

classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "futures==3.1.1",
    "grpcio",
    "grpcio-tools",
    "matplotlib",
    "pytz",
    "pymongo",
    "schedule",
]

[project.optional-dependencies]
develop = [
    "build",
    "bump-my-version",
    "check-manifest",
    "hatchling",
    "pytest",
    "pytest-cov",
    "ruff",
    "twine",
    "wheel",
]

[project.configs]
config = [
    "./config.yml"
]

[project.scripts]

[project.urls]
Repository = "https://github.com/trxe/statsized"
Homepage = "https://github.com/trxe/statsized"

[tool.bumpversion]
current_version = "0.1.0"
commit = true
tag = true

[[tool.bumpversion.files]]
filename = "statsized/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[tool.check-manifest]
ignore = [
    ".copier-answers.yml",
    "Makefile",
    "docs/**/*",
]

[tool.coverage.run]
branch = true
omit = [
    "statsized/tests/integration/",
]
[tool.coverage.report]
exclude_also = [
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@(abc\\.)?abstractmethod",
]
ignore_errors = true
fail_under = 50

[tool.hatch.build]
artifacts = []

[tool.hatch.build.sources]
src = "/"

[tool.hatch.build.targets.sdist]
packages = ["statsized"]

[tool.hatch.build.targets.wheel]
packages = ["statsized"]

[tool.hatch.build.targets.wheel.shared-data]

[tool.pytest.ini_options]
addopts = ["-vvv", "--junitxml=junit.xml"]
asyncio_mode = "strict"
testpaths = "statsized/tests"

[tool.ruff]
line-length = 150

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
combine-as-imports = true
default-section = "third-party"
known-first-party = ["statsized"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
